# apps/ios/Cove/Networking/Generated/openapi.yaml
openapi: "3.1.0"
info:
  title: cove-api
  version: "0.2.0"
  description: |
    Backend-for-frontend gateway for the Cove iOS app.
    Validates Firebase ID tokens and routes traffic to internal services:
      - cove-image  →  /images/*
      - cove-item   →  /discovery, /categories, /items/*, /makers/*, /storefronts/*
      - cove-user   →  /users/me*, /recommendations/*

    The gateway injects the verified Firebase UID as X-Cove-Uid on every
    forwarded request. Downstream services trust this header without
    re-validating the Firebase token.
servers:
  - url: https://api.coveapp.dev
    description: Production (Cloudflare Tunnel)
  - url: https://staging-api.coveapp.dev
    description: Staging (Cloudflare Tunnel)
  - url: http://localhost:8080
    description: Local development

# All routes require a Firebase ID token by default.
# Individual operations can opt out with security: [].
security:
  - bearerAuth: []
paths:
...
  /images/{filename}/url:
    get:
      operationId: getImageURL
      summary: Get a signed imgproxy URL for a stored image
      description: |
        Proxied to cove-image. Returns a short-lived signed URL the iOS client
        can pass directly to AsyncImage. The URL routes through the /i/*
        reverse proxy to imgproxy, which resizes and re-encodes the image on
        first request; Cloudflare caches subsequent transforms at the edge.

        The URL embeds an `exp:` imgproxy processing option and is only valid
        until `expires_at` (default TTL: 1 hour, configurable via
        IMGPROXY_URL_TTL on cove-image).
      parameters:
        - name: filename
          in: path
          required: true
          description: |
            Filename segment of the Garage object key, e.g.
            `5069715ed40e20d8736d6d48ef9f84da8b240e1ead5ba28fdeb110d3d534105c.webp`.
          schema:
            type: string
        - name: w
          in: query
          required: true
          description: Output width in pixels (1–4096).
          schema:
            type: integer
            minimum: 1
            maximum: 4096
        - name: h
          in: query
          required: true
          description: Output height in pixels (1–4096).
          schema:
            type: integer
            minimum: 1
            maximum: 4096
        - name: fit
          in: query
          required: false
          description: |
            Resize mode. `cover` (default) crops to fill the exact dimensions.
            `contain` fits within the dimensions preserving aspect ratio.
          schema:
            type: string
            enum: [cover, contain]
            default: cover
      responses:
        "200":
          description: Signed imgproxy URL generated successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImageURLResponse"
        "400":
          description: Missing or invalid query parameters.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized — Firebase token missing or invalid.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Image not found in Garage.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ── Discovery (cove-item) ─────────────────────────────────────────────────
  /discovery:
    get:
      operationId: getDiscovery
      summary: Discover items
      description: |
        Proxied to cove-item. The core discovery query. Ranks items by a
        blended score:
          score = 0.40 × trust + 0.30 × text_relevance + 0.30 × proximity

        All query parameters are optional. When none are provided, returns all
        active items ranked by trust score descending (max 25 results).

        Proximity scoring requires lat and lon together. radius defaults to
        50 000 m (50 km) when lat/lon are supplied without an explicit value.

        Storefronts with a null location (e.g. online stores) are excluded
        when a geo filter is active but are included in text-only queries.
      parameters:
        - name: q
          in: query
          required: false
          description: Free-text search term (PostgreSQL websearch_to_tsquery).
          schema:
            type: string
        - name: lat
          in: query
          required: false
          description: Latitude in decimal degrees. Must be paired with lon.
          schema:
            type: number
            format: double
        - name: lon
          in: query
          required: false
          description: Longitude in decimal degrees. Must be paired with lat.
          schema:
            type: number
            format: double
        - name: radius
          in: query
          required: false
          description: |
            Search radius in metres. Applies only when lat and lon are present.
            Defaults to 50 000 m when omitted.
          schema:
            type: number
            format: double
        - name: category
          in: query
          required: false
          description: |
            ltree path to scope results to a category subtree.
            Examples: "food.coffee", "food.coffee.whole_bean", "apparel".
          schema:
            type: string
      responses:
        "200":
          description: Discovery results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DiscoveryResponse"
        "400":
          description: Invalid query parameters
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized — Firebase token missing or invalid.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /categories:
    get:
      operationId: getCategories
      summary: Get category tree
      description: |
        Proxied to cove-item. Returns the full category tree (id, name, path,
        children) for the onboarding picker and browse surface. Results are
        ordered by ltree path.
      responses:
        "200":
          description: Category tree
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CategoriesResponse"
        "401":
          description: Unauthorized — Firebase token missing or invalid.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Firebase ID token obtained from the Firebase Auth SDK.
        Attach as "Authorization: Bearer <token>" on all requests except /health.
  schemas:
...
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          description: Human-readable description of what went wrong.
          example: item not found
    ImageURLResponse:
      type: object
      required:
        - url
        - expires_at
      properties:
        url:
          type: string
          description: |
            Signed imgproxy URL valid until expires_at. Pass directly to
            AsyncImage(url:) or URLSession — do not cache beyond expires_at.
          example: "https://api.coveapp.dev/i/abc123.../exp:1748129400/rs:fill:800:800/plain/s3://cove-media/images/5069...webp@webp"
        expires_at:
          type: string
          description: RFC 3339 timestamp after which imgproxy rejects the URL.
          example: "2026-05-26T04:36:37Z"

    # ── Image schemas (cove-item) ─────────────────────────────────────────────
    ImageVariants:
      description: |
        Signed imgproxy URLs for the three list-view variants (thumb, sm, md).
        width and height are the source image dimensions in pixels.
      required: [width, height, thumb, sm, md]
      properties:
        width:
          type: integer
          description: Source image width in pixels.
          example: 800
        height:
          type: integer
          description: Source image height in pixels.
          example: 800
        thumb:
          type: string
          format: uri
          description: 200×200 crop — search result thumbnails, small grid (1×).
        sm:
          type: string
          format: uri
          description: 400×400 crop — list views (2×), avatar-like contexts.
        md:
          type: string
          format: uri
          description: 800×800 crop — detail hero (1×) / list views (3× Pro).
...
    MakerSummary:
      required: [id, name, trust_score]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          example: Onyx Coffee Lab
        trust_score:
          type: number
          example: 3
    StorefrontSummary:
      required: [id, name, type]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        type:
          type: string
          enum: [shop, gallery, studio, market, taproom, online]
          example: online
        distance_m:
          type: number
          description: |
            Distance in metres from the requested location.
            Absent when no geo params were provided.
          example: 1234.5
        listing_url:
          type: string
          format: uri

    # ── Discovery schemas (cove-item) ─────────────────────────────────────────
    DiscoveryResult:
      required: [id, name, maker, nearest_storefront, storefronts, base_trust, score]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          example: Southern Weather Blend
        description:
          type: string
        price_cents:
          type: integer
          example: 2300
        maker:
          $ref: "#/components/schemas/MakerSummary"
        nearest_storefront:
          $ref: "#/components/schemas/StorefrontSummary"
        storefronts:
          type: array
          description: All storefronts that carry this item ("Where to find it").
          items:
            $ref: "#/components/schemas/StorefrontSummary"
        primary_image:
          $ref: "#/components/schemas/ImageVariants"
        base_trust:
          type: number
          description: Maker trust score + nearest storefront trust score.
          example: 3
        score:
          type: number
          description: Blended ranking score (trust + relevance + proximity).
          example: 1.2
    DiscoveryResponse:
      required: [results]
      properties:
        results:
          type: array
          items:
            $ref: "#/components/schemas/DiscoveryResult"

    # ── Category schemas (cove-item) ──────────────────────────────────────────
    CategoryNode:
      required: [id, name, path]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          example: Whole Bean Coffee
        path:
          type: string
          description: ltree path (dot-delimited).
          example: food.coffee.whole_bean
        children:
          type: array
          items:
            $ref: "#/components/schemas/CategoryNode"
    CategoriesResponse:
      required: [categories]
      properties:
        categories:
          type: array
          items:
            $ref: "#/components/schemas/CategoryNode"

    # ── Item detail schemas (cove-item) ───────────────────────────────────────
...